backend/requirements.txt
flask==3.1.3
flask-cors==6.0.2
psycopg[binary]==3.2.12
requests==2.33.1
waitress==3.0.2

# Google Cloud AI Integration
google-generativeai==0.8.3
google-cloud-logging==3.9.0
google-cloud-storage==2.14.0
python-dotenv==1.0.0

# Production server
gunicorn==22.0.0
